Package renamed to NxTransit. Updated docs, tests and readme
