+++ nxtransit/_version.py
@@ -1,0 +1,3 @@
+"""NxTransit package version information."""
+
+__version__ = "0.1.14"
