+++ tests/test_functions.py
@@ -1,4 +1,4 @@
-from transit.functions import last_service
+from nxtransit.functions import last_service
 import networkx as nx
 
 def test_last_service_with_sorted_schedules():
