--- transit/__init__.py
@@ -1,15 +1,0 @@
-"""
-Package for creating multimodal graph of city transit system
-and performing time-depedent calculations using GTFS data
-"""
-VERSION = '0.1.13'
-
-from .routers import time_dependent_dijkstra, single_source_time_dependent_dijkstra
-
-from .loaders import feed_to_graph
-
-from .functions import *
-
-from .converters import parse_seconds_to_time, parse_time_to_seconds
-
-from .connectors import snap_points_to_network
